Kotlin implemtentation of minicap
This is an attempt to write a minicap version in kotlin and
use higher level api that should hopefully make it easier
to support many os version (and the future ones).
It also is likely to prevent the reccurent "Vector<> have
different types" errors.
The implementation is not complete and a few changes will have
to be done on stf.

Signed-off-by: Crepieux Pierre <pierre.crepieux@orange.com>